20714
-Updated tests for the new bulk editor.
-Removed tests that were only needed for the old implementation.
