--- packages/seqdb-ui/components/workflow/pre-library-prep/__tests__/PreLibraryPrepStep.test.tsx
@@ -1,955 +1,0 @@
-import { ResourceSelect } from "common-ui";
-import { PersistedResource } from "kitsu";
-import NumberFormat from "react-number-format";
-import { mountWithAppContext } from "../../../../test-util/mock-app-context";
-import {
-  Chain,
-  ChainStepTemplate,
-  ChainTemplate,
-  Sample,
-  StepResource,
-  StepTemplate
-} from "../../../../types/seqdb-api";
-import { PreLibraryPrepStep } from "../PreLibraryPrepStep";
-
-// Mock out the Link component, which normally fails when used outside of a Next app.
-jest.mock("next/link", () => ({ children }) => <div>{children}</div>);
-
-const TEST_SAMPLES = [
-  { id: "1", type: "sample", name: "test sample 1" },
-  { id: "2", type: "sample", name: "test sample 2" },
-  { id: "3", type: "sample", name: "test sample 3" },
-  { id: "4", type: "sample", name: "test sample 4" },
-  { id: "5", type: "sample", name: "test sample 5" }
-] as PersistedResource<Sample>[];
-
-const TEST_SAMPLE_STEP_RESOURCES: PersistedResource<StepResource>[] = [
-  {
-    id: "1",
-    sample: TEST_SAMPLES[0],
-    type: "stepResource"
-  } as PersistedResource<StepResource>,
-  {
-    id: "2",
-    sample: TEST_SAMPLES[1],
-    type: "stepResource"
-  } as PersistedResource<StepResource>,
-  {
-    id: "3",
-    sample: TEST_SAMPLES[2],
-    type: "stepResource"
-  } as PersistedResource<StepResource>,
-  {
-    id: "4",
-    sample: TEST_SAMPLES[3],
-    type: "stepResource"
-  } as PersistedResource<StepResource>,
-  {
-    id: "5",
-    sample: TEST_SAMPLES[4],
-    type: "stepResource"
-  } as PersistedResource<StepResource>
-];
-
-const TEST_CHAIN_TEMPLATE: PersistedResource<ChainTemplate> = {
-  id: "1",
-  name: "WGS",
-  type: "chainTemplate"
-};
-
-const TEST_CHAIN: PersistedResource<Chain> = {
-  chainTemplate: TEST_CHAIN_TEMPLATE,
-  createdOn: "2019-01-01",
-  id: "1",
-  name: "Mat's chain",
-  type: "chain"
-};
-
-/** This is the first step in the chain that precedes this one. */
-const TEST_SAMPLE_SELECTION_CHAIN_STEP_TEMPLATE: PersistedResource<ChainStepTemplate> = {
-  chainTemplate: TEST_CHAIN_TEMPLATE,
-  id: "1",
-  stepNumber: 1,
-  stepTemplate: { id: "1", type: "stepTemplate" } as PersistedResource<
-    StepTemplate
-  >,
-  type: "chainStepTemplate"
-};
-
-/** This is the second and current step in the chain. */
-const TEST_PRE_LIBRARY_PREP_CHAIN_STEP_TEMPLATE: PersistedResource<ChainStepTemplate> = {
-  chainTemplate: TEST_CHAIN_TEMPLATE,
-  id: "2",
-  stepNumber: 2,
-  stepTemplate: { id: "2", type: "stepTemplate" } as PersistedResource<
-    StepTemplate
-  >,
-  type: "chainStepTemplate"
-};
-
-const TEST_CHAIN_STEP_TEMPLATES = [
-  TEST_SAMPLE_SELECTION_CHAIN_STEP_TEMPLATE,
-  TEST_PRE_LIBRARY_PREP_CHAIN_STEP_TEMPLATE
-];
-
-const mockGet = jest.fn();
-const mockPatch = jest.fn();
-
-const apiContext: any = {
-  apiClient: { get: mockGet, axios: { patch: mockPatch } }
-};
-
-function getWrapper() {
-  return mountWithAppContext(
-    <PreLibraryPrepStep
-      chain={TEST_CHAIN}
-      chainStepTemplates={TEST_CHAIN_STEP_TEMPLATES}
-      step={TEST_PRE_LIBRARY_PREP_CHAIN_STEP_TEMPLATE}
-    />,
-    { apiContext }
-  );
-}
-
-describe("PreLibraryPrepStep UI", () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-
-    /** Mock Kitsu "get" method. */
-    mockGet.mockImplementation(async (path, params) => {
-      if (path === "seqdb-api/stepResource") {
-        if (params.include === "sample") {
-          return { data: TEST_SAMPLE_STEP_RESOURCES };
-        } else if (params.include.includes("sample,preLibraryPrep")) {
-          return { data: [] };
-        }
-      } else {
-        return { data: [] };
-      }
-    });
-
-    /** Mock axios for operations requests. */
-    mockPatch.mockImplementation(async (_, __) => ({ data: [] }));
-  });
-
-  it("Renders the samples from the list", async () => {
-    const wrapper = getWrapper();
-
-    // Await initial queries.
-    await new Promise(setImmediate);
-    wrapper.update();
-
-    expect(mockGet).toHaveBeenCalledTimes(3);
-
-    const [
-      sampleStepResourceCall,
-      ,
-      preLibraryPrepStepResourceCall
-    ] = mockGet.mock.calls;
-
-    expect(sampleStepResourceCall).toEqual([
-      "seqdb-api/stepResource",
-      {
-        // It should be filtered to the previous step's stepResources.
-        filter: {
-          "chain.uuid": "1",
-          "chainStepTemplate.uuid": "1",
-          rsql: ""
-        },
-        include: "sample",
-        page: {
-          limit: 100,
-          offset: 0
-        }
-      }
-    ]);
-    expect(preLibraryPrepStepResourceCall).toEqual([
-      "seqdb-api/stepResource",
-      {
-        fields: {
-          product: "name",
-          protocol: "name",
-          sample: "name,version"
-        },
-        filter: {
-          "chain.uuid": "1",
-          "chainStepTemplate.uuid": "2",
-          rsql: "sample.uuid=in=(1,2,3,4,5)"
-        },
-        include:
-          "sample,preLibraryPrep,preLibraryPrep.protocol,preLibraryPrep.product",
-        page: {
-          limit: 1000
-        }
-      }
-    ]);
-
-    // The tables should show 'not sheared' and 'no size selection'.
-    expect(wrapper.contains("Not Sheared")).toEqual(true);
-    expect(wrapper.contains("No Size Selection")).toEqual(true);
-    expect(wrapper.contains("Sheared")).toEqual(false);
-    expect(wrapper.contains("Size Selection Added")).toEqual(false);
-  });
-
-  it("Lets you add shearing details for the checked samples.", async () => {
-    const wrapper = getWrapper();
-
-    // Await initial queries.
-    await new Promise(setImmediate);
-    wrapper.update();
-
-    // Select samples 3 to 5.
-    wrapper
-      .find(".selected-samples .rt-tbody input[type='checkbox']")
-      .at(2)
-      .prop<any>("onClick")({ target: { checked: true } } as any);
-    wrapper.update();
-    wrapper
-      .find(".selected-samples .rt-tbody input[type='checkbox']")
-      .at(4)
-      .prop<any>("onClick")({
-      shiftKey: true,
-      target: { checked: true }
-    } as any);
-    wrapper.update();
-
-    // Add the preLibraryPrep details.
-    wrapper
-      .find(".inputAmount-field")
-      .find(NumberFormat)
-      .prop<any>("onValueChange")({ floatValue: 1234 } as any);
-    (wrapper
-      .find(".protocol-field")
-      .find(ResourceSelect)
-      .prop("onChange") as any)({ id: "1", type: "protocol" });
-    (wrapper
-      .find(".product-field")
-      .find(ResourceSelect)
-      .prop("onChange") as any)({ id: "1", type: "product" });
-
-    // Assume there are no stepresources for these samples,
-    // so this form submit creates 3 stepResources and does not edit existing ones.
-    mockPatch.mockImplementation(async (_, operations) => {
-      if (operations[0].path === "preLibraryPrep") {
-        return {
-          data: [
-            {
-              data: {
-                attributes: {
-                  preLibraryPrepType: "SHEARING"
-                },
-                id: "1",
-                type: "preLibraryPrep"
-              },
-              status: 201
-            },
-            {
-              data: {
-                attributes: {
-                  preLibraryPrepType: "SHEARING"
-                },
-                id: "2",
-                type: "preLibraryPrep"
-              },
-              status: 201
-            },
-            {
-              data: {
-                attributes: {
-                  preLibraryPrepType: "SHEARING"
-                },
-                id: "3",
-                type: "preLibraryPrep"
-              },
-              status: 201
-            }
-          ]
-        };
-      }
-      if (operations[0].path === "stepResource") {
-        return {
-          data: [
-            { status: 201, data: { id: "11", type: "stepResource" } },
-            { status: 201, data: { id: "12", type: "stepResource" } },
-            { status: 201, data: { id: "13", type: "stepResource" } }
-          ]
-        };
-      }
-    });
-
-    // Submit the form
-    wrapper.find("form.pre-library-prep-form").simulate("submit");
-    // Await form submit.
-    await new Promise(setImmediate);
-
-    // There should have been two patch calls:
-    // One to add the prelibrarypreps
-    // and one to add the stepResources.
-    expect(mockPatch).toHaveBeenCalledTimes(2);
-
-    const [prepCall, stepResourceCall] = mockPatch.mock.calls;
-
-    // There should have been 3 preps created.
-    expect(prepCall).toEqual([
-      "/seqdb-api/operations",
-      [
-        {
-          op: "POST",
-          path: "preLibraryPrep",
-          value: {
-            attributes: {
-              inputAmount: 1234,
-              preLibraryPrepType: "SHEARING"
-            },
-            id: "-100",
-            relationships: {
-              product: {
-                data: {
-                  id: "1",
-                  type: "product"
-                }
-              },
-              protocol: {
-                data: {
-                  id: "1",
-                  type: "protocol"
-                }
-              }
-            },
-            type: "preLibraryPrep"
-          }
-        },
-        {
-          op: "POST",
-          path: "preLibraryPrep",
-          value: {
-            attributes: {
-              inputAmount: 1234,
-              preLibraryPrepType: "SHEARING"
-            },
-            id: "-101",
-            relationships: {
-              product: {
-                data: {
-                  id: "1",
-                  type: "product"
-                }
-              },
-              protocol: {
-                data: {
-                  id: "1",
-                  type: "protocol"
-                }
-              }
-            },
-            type: "preLibraryPrep"
-          }
-        },
-        {
-          op: "POST",
-          path: "preLibraryPrep",
-          value: {
-            attributes: {
-              inputAmount: 1234,
-              preLibraryPrepType: "SHEARING"
-            },
-            id: "-102",
-            relationships: {
-              product: {
-                data: {
-                  id: "1",
-                  type: "product"
-                }
-              },
-              protocol: {
-                data: {
-                  id: "1",
-                  type: "protocol"
-                }
-              }
-            },
-            type: "preLibraryPrep"
-          }
-        }
-      ],
-      expect.anything()
-    ]);
-
-    // There should have been 3 step resources created.
-    expect(stepResourceCall).toEqual([
-      "/seqdb-api/operations",
-      [
-        {
-          op: "POST",
-          path: "stepResource",
-          value: {
-            attributes: {
-              value: "SHEARING"
-            },
-            id: "-100",
-            relationships: {
-              chain: {
-                data: {
-                  id: "1",
-                  type: "chain"
-                }
-              },
-              chainStepTemplate: {
-                data: {
-                  id: "2",
-                  type: "chainStepTemplate"
-                }
-              },
-              preLibraryPrep: {
-                data: {
-                  id: "1",
-                  type: "preLibraryPrep"
-                }
-              },
-              sample: {
-                data: {
-                  id: "3",
-                  type: "sample"
-                }
-              }
-            },
-            type: "stepResource"
-          }
-        },
-        {
-          op: "POST",
-          path: "stepResource",
-          value: {
-            attributes: {
-              value: "SHEARING"
-            },
-            id: "-101",
-            relationships: {
-              chain: {
-                data: {
-                  id: "1",
-                  type: "chain"
-                }
-              },
-              chainStepTemplate: {
-                data: {
-                  id: "2",
-                  type: "chainStepTemplate"
-                }
-              },
-              preLibraryPrep: {
-                data: {
-                  id: "2",
-                  type: "preLibraryPrep"
-                }
-              },
-              sample: {
-                data: {
-                  id: "4",
-                  type: "sample"
-                }
-              }
-            },
-            type: "stepResource"
-          }
-        },
-        {
-          op: "POST",
-          path: "stepResource",
-          value: {
-            attributes: {
-              value: "SHEARING"
-            },
-            id: "-102",
-            relationships: {
-              chain: {
-                data: {
-                  id: "1",
-                  type: "chain"
-                }
-              },
-              chainStepTemplate: {
-                data: {
-                  id: "2",
-                  type: "chainStepTemplate"
-                }
-              },
-              preLibraryPrep: {
-                data: {
-                  id: "3",
-                  type: "preLibraryPrep"
-                }
-              },
-              sample: {
-                data: {
-                  id: "5",
-                  type: "sample"
-                }
-              }
-            },
-            type: "stepResource"
-          }
-        }
-      ],
-      expect.anything()
-    ]);
-  });
-
-  it("Edits existing PreLibraryPreps when they exist, and creates new ones when they don't.", async () => {
-    const wrapper = getWrapper();
-
-    // Await initial queries.
-    await new Promise(setImmediate);
-    wrapper.update();
-
-    // Select samples 3 to 5.
-    wrapper
-      .find(".selected-samples .rt-tbody input[type='checkbox']")
-      .at(2)
-      .prop<any>("onClick")({ target: { checked: true } } as any);
-    wrapper.update();
-    wrapper
-      .find(".selected-samples .rt-tbody input[type='checkbox']")
-      .at(4)
-      .prop<any>("onClick")({
-      shiftKey: true,
-      target: { checked: true }
-    } as any);
-    wrapper.update();
-
-    // Add the preLibraryPrep details.
-    wrapper
-      .find(".inputAmount-field")
-      .find(NumberFormat)
-      .prop<any>("onValueChange")({ floatValue: 4321 } as any);
-
-    mockGet.mockImplementation((path, params) => {
-      if (
-        path === "seqdb-api/stepResource" &&
-        params.include.includes("sample,preLibraryPrep")
-      ) {
-        expect(params.filter.rsql).toEqual("sample.uuid=in=(3,4,5)");
-        // The first 2 samples should already have preLibraryPreps.
-        return {
-          data: [
-            {
-              id: "11",
-              preLibraryPrep: { id: "1", type: "preLibraryPrep" },
-              sample: { id: "3", type: "sample" },
-              type: "stepResource"
-            },
-            {
-              id: "12",
-              preLibraryPrep: { id: "2", type: "preLibraryPrep" },
-              sample: { id: "4", type: "sample" },
-              type: "stepResource"
-            }
-          ]
-        };
-      } else {
-        return { data: [] };
-      }
-    });
-
-    mockPatch.mockImplementation(async (_, operations) => {
-      // The first patch request should edit the first two and create the third.
-      if (operations[0].path === "preLibraryPrep/1") {
-        return {
-          data: [
-            {
-              data: {
-                attributes: {
-                  preLibraryPrepType: "SHEARING"
-                },
-                id: "1",
-                type: "preLibraryPrep"
-              },
-              status: 201
-            },
-            {
-              data: {
-                attributes: {
-                  preLibraryPrepType: "SHEARING"
-                },
-                id: "2",
-                type: "preLibraryPrep"
-              },
-              status: 201
-            },
-            {
-              data: {
-                attributes: {
-                  preLibraryPrepType: "SHEARING"
-                },
-                id: "3",
-                type: "preLibraryPrep"
-              },
-              status: 201
-            }
-          ]
-        };
-      }
-      // The second patch request should create the third stepResource.
-      if (operations[0].path === "stepResource") {
-        return {
-          data: [{ status: 201, data: { id: "13", type: "stepResource" } }]
-        };
-      }
-    });
-
-    // Submit the form
-    wrapper.find("form.pre-library-prep-form").simulate("submit");
-    // Await form submit.
-    await new Promise(setImmediate);
-
-    const [prepCall, stepResourceCall] = mockPatch.mock.calls;
-
-    // There should have been two PATCH operations and one POST operation.
-    expect(prepCall).toEqual([
-      "/seqdb-api/operations",
-      [
-        // The existing ones:
-        {
-          op: "PATCH",
-          path: "preLibraryPrep/1",
-          value: {
-            attributes: {
-              inputAmount: 4321,
-              preLibraryPrepType: "SHEARING"
-            },
-            id: "1",
-            type: "preLibraryPrep"
-          }
-        },
-        {
-          op: "PATCH",
-          path: "preLibraryPrep/2",
-          value: {
-            attributes: {
-              inputAmount: 4321,
-              preLibraryPrepType: "SHEARING"
-            },
-            id: "2",
-            type: "preLibraryPrep"
-          }
-        },
-        // The new one:
-        {
-          op: "POST",
-          path: "preLibraryPrep",
-          value: {
-            attributes: {
-              inputAmount: 4321,
-              preLibraryPrepType: "SHEARING"
-            },
-            id: "-100",
-            type: "preLibraryPrep"
-          }
-        }
-      ],
-      expect.anything()
-    ]);
-
-    // Only one stepResource should have been created: for the new preLibraryPrep
-    expect(stepResourceCall).toEqual([
-      "/seqdb-api/operations",
-      [
-        {
-          op: "POST",
-          path: "stepResource",
-          value: {
-            attributes: {
-              value: "SHEARING"
-            },
-            id: "-100",
-            relationships: {
-              chain: {
-                data: {
-                  id: "1",
-                  type: "chain"
-                }
-              },
-              chainStepTemplate: {
-                data: {
-                  id: "2",
-                  type: "chainStepTemplate"
-                }
-              },
-              preLibraryPrep: {
-                data: {
-                  id: "3",
-                  type: "preLibraryPrep"
-                }
-              },
-              sample: {
-                data: {
-                  id: "5",
-                  type: "sample"
-                }
-              }
-            },
-            type: "stepResource"
-          }
-        }
-      ],
-      expect.anything()
-    ]);
-  });
-
-  it("Does nothing if you submit the form without checking any sample checkboxes.", async () => {
-    const wrapper = getWrapper();
-
-    // Await initial queries.
-    await new Promise(setImmediate);
-    wrapper.update();
-
-    // Submit the form
-    wrapper.find("form.pre-library-prep-form").simulate("submit");
-    // Await form submit.
-    await new Promise(setImmediate);
-
-    // There should have been two empty operations calls.
-    expect(mockPatch.mock.calls).toEqual([
-      ["/seqdb-api/operations", [], expect.anything()],
-      ["/seqdb-api/operations", [], expect.anything()]
-    ]);
-  });
-
-  it("Does nothing if you click the Remove Shearing button without checking any sample checkboxes.", async () => {
-    const wrapper = getWrapper();
-
-    // Await initial queries.
-    await new Promise(setImmediate);
-    wrapper.update();
-
-    // Submit the form
-    wrapper.find("button.remove-shearing").simulate("click");
-    // Await form submit.
-    await new Promise(setImmediate);
-
-    // There should have been one empty operations call.
-    expect(mockPatch.mock.calls).toEqual([
-      ["/seqdb-api/operations", [], expect.anything()]
-    ]);
-  });
-
-  it("Does nothing if you click the Remove Size Selection button without checking any sample checkboxes.", async () => {
-    const wrapper = getWrapper();
-
-    // Await initial queries.
-    await new Promise(setImmediate);
-    wrapper.update();
-
-    // Submit the form
-    wrapper.find("button.remove-size-selection").simulate("click");
-    // Await form submit.
-    await new Promise(setImmediate);
-
-    // There should have been one empty operations call.
-    expect(mockPatch.mock.calls).toEqual([
-      ["/seqdb-api/operations", [], expect.anything()]
-    ]);
-  });
-
-  it("Shows the Shearing and Size Selection status in the table.", async () => {
-    mockGet.mockImplementation(async (path, params) => {
-      // The request for the sample stepResources.
-      if (path === "seqdb-api/stepResource" && params.include === "sample") {
-        // Return stepResources with samples.
-        return {
-          data: [
-            {
-              id: "1",
-              sample: { id: "5", type: "sample" },
-              type: "stepResource"
-            },
-            {
-              id: "2",
-              sample: { id: "6", type: "sample" },
-              type: "stepResource"
-            }
-          ]
-        };
-      }
-
-      // The request for the preLibraryPrep stepResources.
-      if (
-        path === "seqdb-api/stepResource" &&
-        params.include.includes("sample,preLibraryPrep")
-      ) {
-        // Return stepResources with preLibraryPreps and samples.
-        return {
-          data: [
-            {
-              id: "3",
-              preLibraryPrep: {
-                id: "1",
-                preLibraryPrepType: "SHEARING",
-                type: "preLibraryPrep"
-              },
-              sample: { id: "5", type: "sample" },
-              type: "stepResource",
-              value: "SHEARING"
-            },
-            {
-              id: "4",
-              preLibraryPrep: {
-                id: "2",
-                preLibraryPrepType: "SIZE_SELECTION",
-                type: "preLibraryPrep"
-              },
-              sample: { id: "6", type: "sample" },
-              type: "stepResource",
-              value: "SIZE_SELECTION"
-            }
-          ]
-        };
-      }
-
-      // Empty array otherwise so the ResourceSelect doesn't get an undefined response.
-      return { data: [] };
-    });
-
-    const wrapper = getWrapper();
-
-    // Await initial queries.
-    await new Promise(setImmediate);
-    wrapper.update();
-
-    const rows = wrapper.find(".rt-tr");
-
-    // Row 1 should be sheared but not size selected. Row 2 should be size selected but not sheared.
-    expect(rows.at(1).contains("Sheared")).toEqual(true);
-    expect(rows.at(1).contains("No Size Selection")).toEqual(true);
-    expect(rows.at(2).contains("Not Sheared")).toEqual(true);
-    expect(rows.at(2).contains("Size Selection Added")).toEqual(true);
-  });
-
-  it("Lets you delete selected shearing or size selection stepResources.", async () => {
-    mockGet.mockImplementation(async (path, params) => {
-      // Mock sample table response.
-      if (path === "seqdb-api/stepResource" && params.include === "sample") {
-        return {
-          data: [
-            { id: "5", type: "stepResource", sample: { id: "11" } },
-            { id: "6", type: "stepResource", sample: { id: "12" } },
-            { id: "7", type: "stepResource", sample: { id: "13" } },
-            { id: "8", type: "stepResource", sample: { id: "14" } }
-          ]
-        };
-      }
-
-      // Mock stepResource call from inside the deleteStepResources function.
-      if (
-        path === "seqdb-api/stepResource" &&
-        params.include.includes("sample,preLibraryPrep")
-      ) {
-        return {
-          data: [
-            {
-              id: "100",
-              preLibraryPrep: { id: "200", type: "preLibraryPrep" },
-              sample: { id: "12", type: "sample" },
-              type: "stepResource"
-            }
-          ]
-        };
-      }
-
-      return { data: [] };
-    });
-
-    const wrapper = getWrapper();
-
-    // Await initial queries.
-    await new Promise(setImmediate);
-    wrapper.update();
-
-    // Select samples 2 to 4.
-    wrapper
-      .find(".selected-samples .rt-tbody input[type='checkbox']")
-      .at(1)
-      .prop<any>("onClick")({ target: { checked: true } } as any);
-    wrapper.update();
-    wrapper
-      .find(".selected-samples .rt-tbody input[type='checkbox']")
-      .at(3)
-      .prop<any>("onClick")({
-      shiftKey: true,
-      target: { checked: true }
-    } as any);
-    wrapper.update();
-
-    wrapper.find("button.remove-shearing").simulate("click");
-
-    // Await the request and page refresh.
-    await new Promise(setImmediate);
-    wrapper.update();
-
-    expect(mockPatch).toHaveBeenCalledTimes(1);
-    expect(mockPatch).lastCalledWith(
-      "/seqdb-api/operations",
-      [
-        { op: "DELETE", path: "stepResource/100" },
-        { op: "DELETE", path: "preLibraryPrep/200" }
-      ],
-      expect.anything()
-    );
-  });
-
-  it("Shows different view modes for the shearing and size selectino details", async () => {
-    mockGet.mockImplementation(async (path, params) => {
-      // The request for the sample stepResources.
-      if (path === "seqdb-api/stepResource" && params.include === "sample") {
-        return {
-          data: [{ id: "5", type: "stepResource", sample: { id: "10" } }]
-        };
-      }
-
-      // The request for the preLibraryPrep stepResources; There should be a prelibraryprep with
-      // an inputAmount for the sample.
-      if (
-        path === "seqdb-api/stepResource" &&
-        params.include.includes("sample,preLibraryPrep")
-      ) {
-        return {
-          data: [
-            {
-              id: "100",
-              preLibraryPrep: {
-                id: "200",
-                inputAmount: 999,
-                type: "preLibraryPrep"
-              },
-              sample: { id: "10", type: "sample" },
-              type: "stepResource",
-              value: "SIZE_SELECTION"
-            }
-          ]
-        };
-      }
-
-      return { data: [] };
-    });
-
-    const wrapper = getWrapper();
-
-    // Await initial queries.
-    await new Promise(setImmediate);
-    wrapper.update();
-
-    wrapper
-      .find("li.react-tabs__tab.SIZE_SELECTION_DETAILS-toggle")
-      .simulate("click");
-
-    wrapper.update();
-
-    expect(
-      wrapper
-        .find("div.rt-resizable-header-content[children='Input Amount']")
-        .exists()
-    ).toEqual(true);
-
-    // The mock input amount 999 should show up in the table.
-    expect(wrapper.find("div.rt-td[children=999]").exists()).toEqual(true);
-  });
-});
